requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
transformers
google-genai
numpy
